energy_comparison/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dotenv
psycopg2-binary
